Factored `make_tuple` to utils.py.
