+++ namedtensor/torch_helpers.py
@@ -1,17 +1,6 @@
 import torch.nn.functional as F
 from .core import NamedTensorBase, assert_match
-
-
-def make_tuple(names):
-    if names is None:
-        return ()
-    if isinstance(names, tuple):
-        return names
-    if isinstance(names, list):
-        return tuple(names)
-    else:
-        return (names,)
-
+from .utils import make_tuple
 
 class NamedTensor(NamedTensorBase):
     def index_select(self, dim, index):
